move sslkeytool into tools/ and tweak name
